convert Swift4

+++ SwiftPlayer/Source/PlayerCore/Utils/PlayerGLRender.swift
@@ -77,7 +77,7 @@ class MovieGLYUVRender: MovieGLRender {
         }
         
         for i in 0..<3 {
-            glActiveTexture(GLenum(GL_TEXTURE0 + i));
+            glActiveTexture(GLenum(GL_TEXTURE0 + Int32(i)));
             glBindTexture(UInt32(GL_TEXTURE_2D), textures[i]);
             glUniform1i(uniformSamplers[i], GLint(i));
         }
